python/07-staff/requirements.txt
mysql-connector-python
flask
flask-wtf
sqlalchemy
pyecharts
email_validator
